Extract Vault Selector
